Update README.md
The program will be able to generate and store all passwords belonging to an user. 
The user will have the option to specify how man characters they want their password to be.
They can be specify how many characters and the number of specific characters (i.e. 2 symbols, 2 numbers and 1 capital letter).  
They will have the option to store their password, which will be stored by date and what the password is for. 
They can access their saved passwords and can go back and change them at any time. 
The goal is to be as specific as possible.